+++ src/constants/bank.ts
@@ -1,0 +1,61 @@
+/**
+ * 은행 목록 상수 (토스페이 은행 코드 기준)
+ * https://docs-pay.toss.im/guide/bank-code
+ *
+ * 사용처:
+ * - 회원가입 페이지
+ * - 근로자 마이페이지
+ * - 고용주 페이지
+ */
+
+export const BANK_LIST = [
+  // 일반 은행
+  'KDB산업은행',
+  'IBK기업은행',
+  'KB국민은행',
+  'KEB하나은행',
+  '수협은행',
+  'NH농협은행',
+  '우리은행',
+  'SC은행',
+  '씨티은행',
+  '대구은행',
+  '부산은행',
+  '광주은행',
+  '제주은행',
+  '전북은행',
+  '경남은행',
+  'MG새마을금고',
+  '신협',
+  '저축은행',
+  '산림조합',
+  '우체국',
+  '하나은행',
+  '신한은행',
+  '케이뱅크',
+  '카카오뱅크',
+  '토스뱅크',
+  'SBI저축은행',
+  // 증권사
+  'KB증권',
+  '미래에셋증권',
+  '삼성증권',
+  '한국투자증권',
+  'NH투자증권',
+  '교보증권',
+  '하이투자증권',
+  '현대차투자증권',
+  '키움증권',
+  '이베스트증권',
+  'SK증권',
+  '대신증권',
+  '한화투자증권',
+  '하나증권',
+  '토스증권',
+  '신한투자증권',
+  'DB금융투자',
+  '유진투자증권',
+  '메리츠증권',
+] as const;
+
+export type BankName = (typeof BANK_LIST)[number];
